fix script for support sqlit and pg buildbot in a server

+++ proteus/install_buildbot_slave_env.py
@@ -1,14 +1,16 @@
 from fabric.context_managers import prefix
+from fabric.contrib.files import exists
 from fabric.operations import run, sudo
 from profab.role import Role
 
 def install_buildbot_slave_env(server, virtenv_path):
-    sudo('easy_install pip')
-    sudo('easy_install virtualenv')
-    sudo("virtualenv --no-site-packages %s" % (virtenv_path), user="www-data")
-    with prefix("source %s/bin/activate" % (virtenv_path)):
-        sudo("pip install twisted==12.0.0", user="www-data")
-        sudo("pip install buildbot-slave==0.8.6", user="www-data")
+    if not exists(virtenv_path):
+        sudo('easy_install pip')
+        sudo('easy_install virtualenv')
+        sudo("virtualenv --no-site-packages %s" % (virtenv_path), user="www-data")
+        with prefix("source %s/bin/activate" % (virtenv_path)):
+            sudo("pip install twisted==12.0.0", user="www-data")
+            sudo("pip install buildbot-slave==0.8.6", user="www-data")
  
 class Configure(Role):
     """
